Add files via upload
In the hope of getting to know more about Mars and the possibility of life on its surface, a huge number of 
exploratory missions to different regions of the planet need to be conducted. Imagine (at some future time) 
that we have several rovers of different types and capabilities on the surface of Mars ready to carry out 
missions to its different regions. There also exists a hypothetical base station which acts as the central point 
from which the different rovers begin any of their exploratory missions and to which they return after the 
mission completion. If we suppose that new missions are formulated regularly, we then need to assign those 
new missions to the different rovers that we have available. Using your programming skills and your 
knowledge of the different data structures, you are going to develop a program that simulates the mission 
assignment process and calculates some related statistics in order to help improve the overall process.